codigo_fonte/requirements.txt
cffi==1.17.1
chardet==5.2.0
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
psutil==6.1.0
pycparser==2.22
pywin32==308
reportlab==4.2.5
sounddevice==0.5.1
WMI==1.5.1
